+++ CHANGELOG.md
@@ -14,11 +14,12 @@ This changelog was started for release 0.0.3.
 - empty_ok_if key for validator & templates
 - empty_ok_unless key for validator & templates
 - readme key for validator
-- unique key for validator 
+- unique key for validator
 - expected_rows key for templates
 - logs parameters for templates
 - na_ok key for validators & templates
-- generate_validation key for validators & templates
+- skip_generation key for validators & templates
+- skip_validation key for validators & templates
 
 ### Fixed